Refactor: 메세지 카드 렌더링 하는 로직 컴포넌트로 분리
- 메세지 모달도 같은 컴포넌트에 분리
- MessageCardContents 이름 변경
